--- backend/src/main/java/com/backend/member/MemberService.java
@@ -1,15 +1,0 @@
-package com.backend.member;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface MemberService {
-    boolean insertMember(Member member);
-
-
-    boolean existMemberId(String memberId);
-
-    boolean existEmail(String email);
-
-    boolean existPhone(String phone);
-}
