+++ backend/src/main/java/com/backend/member/entity/Member.java
@@ -2,8 +2,13 @@ package com.backend.member.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Getter
@@ -28,11 +33,23 @@ public class Member {
     @Column(name = "member_phone", unique = true)
     private String phone;
 
-    @Column(name = "member_birth")
-    private String birth;
+    @Column(name = "member_BirthDate")
+    private String birthDate;
 
     @Column(name = "member_image")
     private String image;
+
+    @Column(name="member_SignUpTime")
+    private LocalDate memberSignUpTime = LocalDate.now();
+
+    @Column(name="member_Info_Yn", nullable = false)
+    private boolean infoAgree;
+
+    @Column(name="member_Sns_Push_Yn")
+    private boolean snsPushYn;
+
+    @Column(name="member_Email_Push_Yn")
+    private boolean emailPushYn;
 
     @ElementCollection(fetch = FetchType.EAGER)
     @Column(name = "member_role")
@@ -53,13 +70,13 @@ public class Member {
 //    private Set<Authority> authorities;
 
     @Builder
-    public Member(String id, String name, String password, String email, String phone, String birth, String image, List<String> role) {
+    public Member(String id, String name, String password, String email, String phone, String birthDate, String image, List<String> role) {
         this.id = id;
         this.name = name;
         this.password = password;
         this.email = email;
         this.phone = phone;
-        this.birth = birth;
+        this.birthDate = birthDate;
         this.image = image;
         this.role = role;
     }
@@ -68,5 +85,7 @@ public class Member {
         this.password = password;
     }
 
-
+    public void setPassword(String password, PasswordEncoder passwordEncoder) {
+        this.password = passwordEncoder.encode(password);
+    }
 }
