+++ backend/src/main/java/com/backend/member/repository/MemberRepository.java
@@ -14,4 +14,8 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByNameAndIdAndEmail(String name, String id, String email);
 
     Optional<Member> deleteById(String id);
+
+    boolean existsById(String id);
+    boolean existsByEmail(String email);
+    boolean existsByPhone(String phone);
 }
