+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import "./styles/Reset.css";
-import RoutesTest from "./routes/RoutesTest";
+import Routes from "./routes/Routes";
 import Header from "./components/common/Header";
 import Footer from "./components/common/Footer";
 import Notice from "./components/common/Notice";
@@ -14,7 +14,7 @@ function App() {
       <Header />
       {/* lazy() 사용시 서스펜스 적용 안하면 오류발생함 */}
       <Suspense fallback={<Loading />}>
-        <RoutesTest></RoutesTest>
+        <Routes></Routes>
       </Suspense>
       <Footer />
     </Router>
