+++ frontend/src/components/common/NotFound.js
@@ -1,7 +1,7 @@
 import React from "react";
-import error from "../assets/404error.json";
+import error from "../../assets/404error.json";
 import Lottie from "react-lottie";
-import "../styles/NotFound.css";
+import "../../styles/NotFound.css";
 import { Link } from "react-router-dom";
 
 const NotFound = () => {
