--- frontend/src/constants/Messages.js
@@ -1,19 +1,0 @@
-export const CONFIRM_MESSAGE = {
-  DELETE: "정말 삭제하시겠습니까?",
-  LOGOUT: "로그아웃하시겠습니까?",
-  WITHDRAWAL: "정말 탈퇴하시겠습니까?",
-};
-
-export const SIGNUP_ERROR_MESSAGE = {
-  ID: "id가 5자 이상, 15자 이하의 문자열이며 대소문자/숫자인지",
-  PW: "입력된 비밀번호 값이 8자 이상이며, 영문 대/소문자와 숫자, 그리고 특수문자가 각각 하나 이상 포함되어 있는지.",
-  CONFIRMPW: "그리고 비밀번호와 확인 비밀번호가 일치하는지",
-  PHONE: "11자리의 숫자만 입력하였는지",
-  NAME: "입력된 이름 값이 비어있지 않은 문자열이며 2-4글자의 한글이름인지",
-  EMAIL: "입력된 이메일 값이 이메일 주소의 형식에 맞는지",
-  INFOAGREE: "정보동의란 동의 했는지",
-  PHONE_CERTIFICATE_WRONG: "인증번호가 틀렸다.",
-  ID_DUPLICATE: "중복된 ID가 존재한다",
-  PHONE_DUPLICATE: "중복된 전화번호가 존재한다",
-  EMAIL_DUPLICATE: "중복된 이메일이가 존재한다",
-};
